hub_application/requirements.txt
# psycopg2
psycopg2-binary==2.9.6
numpy~=1.24.1
opencv_contrib_python==4.7.0.72
opencv-python~=4.7.0.72
Pillow~=9.3.0
torch~=2.0.1
torchvision~=0.15.2
tqdm~=4.64.1
tensorboardX~=2.6.1
easydict~=1.10
pip~=23.1.2
cryptography~=39.0.0
lxml~=4.9.2
tornado~=6.2
setuptools~=65.5.0
pytz~=2022.7.1
cffi~=1.15.1
cvzone~=1.5.6
mediapipe~=0.10.2
imutils~=0.5.4
scipy~=1.9.3
APScheduler~=3.10.4
Werkzeug==2.3.7
paho-mqtt~=1.6.1
requests~=2.31.0
dlib~=19.24.2
face_recognition==1.3.0